Massive feature expansion: ~300 new source files adding ~404 missing AE features
Core Tools & Edit Operations:
- Pen tool, shape tools, freehand, selection, brush, camera, puppet, transform
- Pre-compose, split layer, freeze frame, auto-trace, shapes from text

3D Features:
- OBJ/FBX/C4D model importers, C4D-style software renderer
- 3D camera tracker (Harris corners, KLT flow, RANSAC)
- Full material options, camera options, layer switches

Effects Suite (200+ total):
- 15 categories: blur, color correction, distort, generate, keying, etc.
- Audio effects: EQ, compressor, limiter, reverb, delay, pitch shift
- All with real pixel-level rendering

Expression Engine:
- Full JavaScript lexer/parser/evaluator, all AE functions
- Wiggle, loopOut, linear, ease, clamp, vector math, layer refs

Scripting:
- ExtendScript engine with Math.*, Project/Comp/Layer API

Performance:
- GPU renderer with CPU fallback, disk cache, profiler
- Thread pool, spatial grid, frame cache

Import/Export:
- EXR, HDR, DPX, TGA, PSD, ProRes, DNxHD, MXF, image sequences
- OCIO/ACES color management, 3D LUT, ICC profiles

Collaboration:
- Team project versioning, asset library, auto-save
- Project manager, footage interpreter, solids library

UI Panels:
- 25 total panels, all menus, keyboard shortcuts
- Viewport: multi-view, channel display, safe margins, ROI

All 66 tests passing, CI with 6 optional feature flags

+++ core/effects/noise/add_grain_effect.cpp
@@ -37,13 +37,13 @@ void AddGrainEffect::render(PixelBuffer& buffer, double time) {
             uint8_t* p = buffer.pixelAt(x, y);
             if (mono) {
                 double n = dist(gen) * scale;
-                p[0] = static_cast<uint8_t>(std::clamp(p[0] + n, 0.0, 255.0));
-                p[1] = static_cast<uint8_t>(std::clamp(p[1] + n, 0.0, 255.0));
-                p[2] = static_cast<uint8_t>(std::clamp(p[2] + n, 0.0, 255.0));
+                p[0] = static_cast<uint8_t>(std::clamp(static_cast<double>(p[0] + n), 0.0, 255.0));
+                p[1] = static_cast<uint8_t>(std::clamp(static_cast<double>(p[1] + n), 0.0, 255.0));
+                p[2] = static_cast<uint8_t>(std::clamp(static_cast<double>(p[2] + n), 0.0, 255.0));
             } else {
-                p[0] = static_cast<uint8_t>(std::clamp(p[0] + dist(gen) * scale, 0.0, 255.0));
-                p[1] = static_cast<uint8_t>(std::clamp(p[1] + dist(gen) * scale, 0.0, 255.0));
-                p[2] = static_cast<uint8_t>(std::clamp(p[2] + dist(gen) * scale, 0.0, 255.0));
+                p[0] = static_cast<uint8_t>(std::clamp(static_cast<double>(p[0] + dist(gen) * scale), 0.0, 255.0));
+                p[1] = static_cast<uint8_t>(std::clamp(static_cast<double>(p[1] + dist(gen) * scale), 0.0, 255.0));
+                p[2] = static_cast<uint8_t>(std::clamp(static_cast<double>(p[2] + dist(gen) * scale), 0.0, 255.0));
             }
         }
     }
